Refactor submission preprocessing into background workers
